Compress brushpolyvert_t to 32 bytes

+++ Renderer/r_model.h
@@ -91,7 +91,7 @@ typedef struct msurface_s {
 	// lighting info
 	int			dlightframe;
 
-	int			lightmaptexturenum;
+	unsigned short		lightmaptexturenum;
 	byte		styles[MAXLIGHTMAPS];
 	byte		*samples;		// [numstyles * surfsize]
 } msurface_t;
